Collapsing Visitable into Document
(cherry picked from commit ca2ca35e6466991f470f3e2f5fa7dbc3d1dc913f)
